pyproject.toml
[project]
name = "class-recorder"
version = "1.0.0"
description = "Record, transcribe, and summarize lectures for Obsidian"
requires-python = ">=3.9"

dependencies = [
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "openai>=1.0.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "tenacity>=8.2.0",
]

[project.scripts]
recorder = "class_recorder.cli:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["class_recorder*"]
exclude = ["recordings*", "temp*"]
